fix: importar funcionalidades relacionadas a alterar turno para classes de refeição, para evitar redundancia
