+++ saru-rest/src/main/java/saru/saru_rest/controller/RefeicaoController.java
@@ -8,8 +8,7 @@ import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.bind.annotation.*;
 import saru.saru_rest.dtos.RefeicaoDTO;
 import saru.saru_rest.entity.RefeicaoEntity;
-import saru.saru_rest.exceptions.RefeicaoJaCompradaException;
-import saru.saru_rest.exceptions.SaldoInsuficienteException;
+import saru.saru_rest.exceptions.*;
 import saru.saru_rest.repository.RefeicaoRepository;
 import saru.saru_rest.service.QRCodeService.QRCodeService;
 import saru.saru_rest.service.refeicao.RefeicaoService;
@@ -46,4 +45,9 @@ public class RefeicaoController {
                 .body(qrCodeService.getQRCodeImage(refeicaoEntity));
     }
 
+    @PutMapping(value="/alterarTurno")
+    public ResponseEntity<String> alterarTurno(@RequestBody RefeicaoDTO refeicaoDTO) throws TodasRefeicoesCompradasException, TurnoJaCompradoException, SemRefeicoesCompradasException {
+        String cpf = (String) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        return ResponseEntity.ok().body(refeicaoService.alterarTurno(cpf, refeicaoDTO.getTurno()));
+    }
 }
