+++ saru-rest/src/main/java/saru/saru_rest/service/refeicao/RefeicaoServiceImpl.java
@@ -3,11 +3,13 @@ package saru.saru_rest.service.refeicao;
 import org.springframework.stereotype.Service;
 import saru.saru_rest.dtos.RefeicaoDTO;
 import saru.saru_rest.entity.RefeicaoEntity;
-import saru.saru_rest.exceptions.RefeicaoJaCompradaException;
-import saru.saru_rest.exceptions.SaldoInsuficienteException;
+import saru.saru_rest.entity.enums.Turno;
+import saru.saru_rest.exceptions.*;
 import saru.saru_rest.repository.ClienteRepository;
 import saru.saru_rest.repository.RefeicaoRepository;
 import saru.saru_rest.service.QRCodeService.QRCodeService;
+
+import java.util.List;
 
 @Service
 
@@ -48,6 +50,38 @@ public class RefeicaoServiceImpl implements RefeicaoService {
 
     }
 
+    public String alterarTurno(String cpf, Turno turno) throws SemRefeicoesCompradasException, TodasRefeicoesCompradasException, TurnoJaCompradoException {
+        List<RefeicaoEntity> refeicoes= refeicaoRepository.findByCpfCliente(cpf);
+        try{
+            if (clienteRepository.existsById(cpf) && verificaTurno(refeicoes, turno) && verificaSemRefeicoesCompradas(refeicoes) && verificaTodasRefeicoesCompradas(refeicoes)) {
+                refeicoes.getFirst().setTurno(turno);
+                refeicoes.getFirst().generateNewToken();
+                refeicaoRepository.save(refeicoes.getFirst());
 
+                return "Turno alterado com sucesso";
+            }
+        }catch (Exception e){
+            e.printStackTrace();
+        }
+        return "Erro ao alterar turno";
+    }
+    public boolean verificaTurno(List<RefeicaoEntity> refeicoes, Turno turno) throws TurnoJaCompradoException{
+        if (refeicoes.get(0).getTurno().equals(turno)){
+            throw new TurnoJaCompradoException();
+        }
+        return true;
+    }
+    public boolean verificaTodasRefeicoesCompradas(List<RefeicaoEntity> refeicoes) throws TodasRefeicoesCompradasException {
+        if (refeicoes.size() >= 2){
+            throw new TodasRefeicoesCompradasException();
+        }
+        return true;
+    }
+    public boolean verificaSemRefeicoesCompradas(List<RefeicaoEntity> refeicoes) throws SemRefeicoesCompradasException {
+        if (refeicoes.isEmpty()) {
+            throw new SemRefeicoesCompradasException();
+        }
+        return true;
+    }
 
 }
